requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.10.0
django-phone-field==1.8.1
Pillow==9.4.0
pytz==2020.4
pyzbar==0.1.8
qrcode==6.1
segno==1.3.3
sqlparse==0.4.1
numpy==1.24.1
opencv-python==4.7.0.68
twilio==7.9.2
stripe==3.5.0
